remove writable trait for WriteBatch
